Revert "feat: adapt to Windows platform"
