+++ net_dialer.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !windows
-// +build !windows
-
 package netpoll
 
 import (
@@ -22,6 +19,15 @@ import (
 	"net"
 	"time"
 )
+
+// Dialer extends net.Dialer's API, just for interface compatibility.
+// DialConnection is recommended, but of course all functions are practically the same.
+// The returned net.Conn can be directly asserted as Connection if error is nil.
+type Dialer interface {
+	DialConnection(network, address string, timeout time.Duration) (connection Connection, err error)
+
+	DialTimeout(network, address string, timeout time.Duration) (conn net.Conn, err error)
+}
 
 // DialConnection is a default implementation of Dialer.
 func DialConnection(network, address string, timeout time.Duration) (connection Connection, err error) {
