+++ sys_epoll_linux.go
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build !arm64
 // +build !arm64
 
 package netpoll
