+++ sys_keepalive_unix.go
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build netbsd || freebsd || dragonfly || linux
 // +build netbsd freebsd dragonfly linux
 
 package netpoll
